Fix deleted contact bug
Added "deleted = 0" to WHERE clause to prevent outputting
deleted contacts.

Also fixed bug for IM fields not being written with Android 2.3
Gingerbread contacts2.db databases. This was due to data.data5
not being used to identify IM service in that version. Now
handled with a default writing the IM field only to the top of
the NOTE field.